src/test/java/lotto/domain/LottoTest.java
package lotto.domain;

import lotto.view.message.exception.LottoGenerateErrorMessage;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;

class LottoTest {
    @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
    @Test
    void createLottoByOverSize() {
        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage(LottoGenerateErrorMessage.INCORRECT_SIZE.getMessage());
    }

    @DisplayName("로또 번호의 개수가 6개보다 적으면 예외가 발생한다.")
    @Test
    void createLottoByUnderSize() {
        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4)))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage(LottoGenerateErrorMessage.INCORRECT_SIZE.getMessage());
    }

    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
    @Test
    void createLottoByDuplicatedNumber() {
        // TODO: 이 테스트가 통과할 수 있게 구현 코드 작성
        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage(LottoGenerateErrorMessage.DUPLICATED_NUMBER.getMessage());
    }

    @DisplayName("로또 번호에 1~45 범위를 벗어난 숫자가 있으면 예외가 발생한다.")
    @Test
    void createLottoByOutOfRangeNumber() {
        assertThatThrownBy(() -> new Lotto(List.of(0, 2, 3, 4, 5, 6)))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage(LottoGenerateErrorMessage.OUT_OF_RANGE_NUMBER.getMessage());

        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 60)))
                .isInstanceOf(IllegalArgumentException.class)
                .hasMessage(LottoGenerateErrorMessage.OUT_OF_RANGE_NUMBER.getMessage());
    }

    @DisplayName("로또 번호는 오름차순으로 정렬되어 저장되어야 한다.")
    @Test
    void createLottoByUnorderedNumbers() {
        List<Integer> case1 = List.of(9, 1, 41, 15, 38, 24);

        Lotto result1 = new Lotto(case1);

        assertThat(result1.getNumbers()).isEqualTo(List.of(1, 9, 15, 24, 38, 41));
    }
}